Run perl script and manually fix some issues in src folder.

+++ src/piqmee/util/AttachmentTimesAll.java
@@ -1,7 +1,11 @@
 package piqmee.util;
 
-import beast.core.*;
-import beast.evolution.tree.Node;
+import beast.base.core.Description;
+import beast.base.core.Function;
+import beast.base.core.Input;
+import beast.base.core.Loggable;
+import beast.base.evolution.tree.Node;
+import beast.base.inference.CalculationNode;
 import piqmee.tree.QuasiSpeciesTree;
 import piqmee.tree.QuasiSpeciesNode;
 
@@ -20,7 +24,8 @@ public class AttachmentTimesAll extends CalculationNode implements Function, Log
 
     private QuasiSpeciesTree qsTree;
 
-    public AttachmentTimesAll() { };
+    public AttachmentTimesAll() {
+    };
 
     @Override
     public void initAndValidate() {
@@ -30,7 +35,7 @@ public class AttachmentTimesAll extends CalculationNode implements Function, Log
     @Override
     public int getDimension() {
         int size = 0;
-        for (Node node : qsTree.getExternalNodes()){
+        for (Node node : qsTree.getExternalNodes()) {
             size += ((QuasiSpeciesNode) node).getAttachmentTimesList().length;
         }
         return size;
@@ -41,7 +46,7 @@ public class AttachmentTimesAll extends CalculationNode implements Function, Log
         if (iDim < getDimension()) {
             int index = iDim;
             int size;
-            for (Node node : qsTree.getExternalNodes()){
+            for (Node node : qsTree.getExternalNodes()) {
                 size = ((QuasiSpeciesNode) node).getAttachmentTimesList().length;
                 if (index < size) {
                     return ((QuasiSpeciesNode) node).getAttachmentTimesList()[index];
@@ -55,11 +60,11 @@ public class AttachmentTimesAll extends CalculationNode implements Function, Log
     }
 
     @Override
-    public void init(PrintStream out){
+    public void init(PrintStream out) {
 
         String idString = qsTree.getID();
         // print all haplo names
-        for (Node node : qsTree.getExternalNodes()){
+        for (Node node : qsTree.getExternalNodes()) {
             int maxtime = ((QuasiSpeciesNode) node).getAttachmentTimesList().length;
             for (int time = 0; time < maxtime; time++) {
                 String haploname = node.getID();
@@ -69,7 +74,7 @@ public class AttachmentTimesAll extends CalculationNode implements Function, Log
     }
 
     @Override
-    public void log(int nSample, PrintStream out) {
+    public void log(long nSample, PrintStream out) {
         // print all haplo names
         for (Node node : qsTree.getExternalNodes()) {
             double[] attachmentTimes = ((QuasiSpeciesNode) node).getAttachmentTimesList();
@@ -80,5 +85,6 @@ public class AttachmentTimesAll extends CalculationNode implements Function, Log
     }
 
     @Override
-    public void close(PrintStream out) { }
+    public void close(PrintStream out) {
+    }
 }
